Move breakpoint stuff to breakpoint.go
